2021/april/separating grpc services (#244)
The goal of this PR is to split the gRPC service into two services that can evolve separately.

The OptimizerService now has a reduced surface and in subsequent PRs we will evolve it further. The goal is to make it easy to implement clients in different languages, and to reduce the cost of maintaining a public-facing API.

The OptimizerMonitoringService contains many of the interfaces likely to evolve faster. In general we shouldn't need to worry about backwards compatibility of this service as it is meant to be shipped with its client (the Tomograph, etc.).

Note, the goal of this PR is to simply separate the two services - no new functionality was added, nor was any removed. Thus existing tests should cover these changes.

Implementation

Splitting the services required creating a second .proto file and specifying the OptimizerMonitoringService there. There is some code duplication, which is the price we pay for decoupling of the two services. This duplication should be reduced as the two services evolve their separate ways.

In order to properly serialize/deserialize the OptimizationProblem, I wrote the OptimizerServiceEncoder/Decoder classes and OptimizerMonitoringServiceEncoder/Decoder classes. These will be further developed to encode/decode other objects too, as we move away from the JSON strings inside protobuf messages.

+++ source/Mlos.Python/mlos/Optimizers/ExperimentDesigner/UtilityFunctionOptimizers/RandomNearIncumbentOptimizer.py
@@ -212,7 +212,7 @@ class RandomNearIncumbentOptimizer(UtilityFunctionOptimizer):
             self.logger.info(error_message)
             raise UnableToProduceGuidedSuggestionException(error_message)
 
-        if incumbents_df.dtypes['utility'] != np.float:
+        if incumbents_df.dtypes['utility'] != float:
             self.logger.info(
                 f"The type of incumbents_df['utility'] is {incumbents_df.dtypes['utility']}. Utility function: {self.utility_function.__class__.__name__}, "
                 f"incumbents_df length: {len(incumbents_df.index)}"
@@ -402,7 +402,7 @@ class RandomNearIncumbentOptimizer(UtilityFunctionOptimizer):
             product=True
         )
         utility_df = self.utility_function(feature_values_pandas_frame=features_df)
-        assert utility_df.dtypes['utility'] == np.float,\
+        assert utility_df.dtypes['utility'] == float,\
             f"{utility_df} produced by {self.utility_function.__class__.__name__} has the wrong type for the 'utility' column: {utility_df.dtypes['utility']}"
         return utility_df
